ESTS-137452 Upgrade to 2.30.0 and remove patch
It seems that the RackHD 2.29.0 docker images got rebuilt from master
on Nov 3 even though 2.29.0 was tagged on Oct 30. Thus the 2.29.0
images now contain a couple of that patches that this RPM attempts to
apply that were merged into master after 2.29.0 was released. That
caused the patches to start failing.

To fix this just went ahead and upgraded to RackHD 2.30.0 and removed
the offending patches.